% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Bin2MultiSV}
\alias{Bin2MultiSV}
\title{Read data using configuration file}
\usage{
Bin2MultiSV(CfgFile)
}
\arguments{
\item{CfgFile}{: A tab separated file with the following format (with no header)


Contrast            SampleID            PathtoRDbinFile


Treat                Sample1                ~/Sample1.WindowRD-1000.count}
}
\value{
\code{Bin2MultiSV}: read configuaration file
}
\description{
Read data using configuration file
}
\examples{
\dontrun{
ConfigFile <- "~/MultiSV.conf
Bin2MultiSV(ConfigFile)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
Bin2MultiSV
}

