% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CallMultiSV}
\alias{CallMultiSV}
\title{A Method to identify MultiSV from RDbin files for different samples defined in Configuration File}
\usage{
CallMultiSV(CfgFile, LgLim, SVSize, MtlSVOut)
}
\arguments{
\item{CfgFile}{: A tab separated file with the following format (with no header)


  Contrast  SampleID     PathtoRDbinFile


Treat    Sample1    ~/Sample1.WindowRD-1000.count


Control    Ref1    ~/Ref1.WindowRD-1000.count

Contrast    SampleID    PathtoRDbinFile}

\item{LgLim}{: log2 limit, set 0.6 or higher}

\item{SVSize}{: Minimum size of SV to identify, set 5000 or higher}

\item{MtlSVOut}{: Output file}
}
\value{
\code{CallMultiSV}:
}
\description{
This is the main function to identify SVs. It works after the data has been prepared using
Bin2MultiSV from bam files. Bin2MultiSV will prepare the data for MultiSV package and  CallMultiSV
will identify SVs.
}
\details{
Data Input from read depth (RD) bins
}
\examples{
\dontrun{
CallMultiSV(CfgFile, LgLim, SVSize, MtlSVOut)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
Bin2MultiSV
}

