% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CmptMltPvl}
\alias{CmptMltPvl}
\title{Compute Pvals of MultiSVs}
\usage{
CmptMltPvl(MultiData,CfgFile)
}
\arguments{
\item{CfgFile}{: A tab separated file with the following format (with no header)


    Contrast            SampleID            PathtoRDbinFile


    Treat                Sample1                ~/Sample1.WindowRD-1000.count}

\item{MultiData}{: MultiSV Data object}
}
\value{
\code{CmptMltPvl}: 0.05
}
\description{
Compute Pvals of MultiSVs
}
\examples{
\dontrun{
CmptMltPvl("MultiSV.conf")
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
CmptMltPvl
}

