% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetTestCfgFile}
\alias{GetTestCfgFile}
\title{Function with MultiSV input example}
\usage{
GetTestCfgFile()
}
\description{
Function with MultiSV input example
}
\examples{
##' @export
`MultiSVExample.default` <- function(MultiSVData){
  LgLim = 0.6
  SVSize = 5000
  MultiData <- PrepMultiDt(MultiSVData)
  MultiData <- PrepMultiDtLgMn(MultiData)
  MultiData <- ProcMutilDt(MultiData)
  IdfMltSV(MultiData,LgLim,SVSize)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}

