% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gtExEs}
\alias{gtExEs}
\title{proc information}
\usage{
gtExEs(proc)
}
\arguments{
\item{proc}{: Name of proc}
}
\value{
\code{gtExEs}: proc
}
\description{
proc information
}
\examples{
\dontrun{
gtExEs(proc)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
Bin2MultiSV
}

