% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IdfMltSV}
\alias{IdfMltSV}
\title{Call MultiSVs}
\usage{
IdfMltSV(MultiData,LgLim,SVSize)
}
\arguments{
\item{MultiData}{: MultiSV Data object}

\item{LgLim}{: log2 threshold}

\item{SVSize}{: Minimum size of SV to identify, set 5000 or higher}
}
\value{
\code{IdfMltSV}: read configuaration file
}
\description{
Call MultiSVs
}
\examples{
\dontrun{
IdfMltSV(MultiData,Cfg,LgLim,SVSize)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}

