% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MultiSVData}
\alias{MultiSVData}
\title{RD Data}
\description{
The data set is read depth (RD) from six domestic populations including ugc355, ugc356,
ugc357, ugc358, ugc376, ugc377 and one wild (WB) population of pigs (SS10.2).
The reads for the populations are available from SRA accession number SRA057461.
RD from each population is in 1kb bins obtained after aligning the reads
from each population to SS10.2 reference genome assembly.
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\references{
Rubin, C.-J., H.-J.Megens, A.Martinez Barrio, K.Maqbool, S. Sayyab,
D. Schwochow, C.Wang, Ö. Carlborg, P. Jern, C.B. Jørgensen, A.
L.Archibald, M. Fredholm, M. A. M. Groenen, and L.Andersson.
2012. Strong signatures of selection in the domestic pig genome.
Proc. Natl. Acad. Sci. USA 109:19529–19536, DOI: 10.1073/pnas.1217149109
}
\keyword{CNV}
\keyword{Deletion}
\keyword{Duplication}
\keyword{MultiSV}
\keyword{Population}

