% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MultiSVExample}
\alias{MultiSVExample}
\title{MultiSV Example}
\usage{
MultiSVExample(MultiSVData)
}
\arguments{
\item{MultiSVData}{: MultiSV Data object}
}
\value{
\code{MultiSVExample}: Run a small example to identify MultiSV
}
\description{
MultiSV Example
}
\examples{
\dontrun{
MultiSVExample(MultiSVData)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
CallMultiSV
}

