% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrbMlt}
\alias{PrbMlt}
\title{Fit linear mixed model}
\usage{
PrbMlt(MultiData)
}
\arguments{
\item{MultiData}{: MultiSV Data object}
}
\value{
\code{PrbMlt}: read configuaration file
}
\description{
Fit linear mixed model
}
\examples{
\dontrun{
PrbMlt(MultiData)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
PrbMlt
}

