% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrepMultiDt}
\alias{PrepMultiDt}
\title{Prepare MultiSV data}
\usage{
PrepMultiDt(MultiData)
}
\arguments{
\item{MultiData}{: MultiSV Data object}
}
\value{
\code{PrepMultiDt}: Prepare MultiSV data
}
\description{
Prepare MultiSV data
}
\examples{
\dontrun{
PrepMultiDt(MultiData)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
PrepMultiDt
}

