% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrepMultiDtLgMn}
\alias{PrepMultiDtLgMn}
\title{Prepare MultiSV data}
\usage{
PrepMultiDtLgMn(MultiData)
}
\arguments{
\item{MultiData}{: MultiSV Data object}
}
\value{
\code{PrepMultiDtLgMn}: Prepare MultiSV data
}
\description{
Prepare MultiSV data
}
\examples{
\dontrun{
PrepMultiDtLgMn(MultiData)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
PrepMultiDt
}

