% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProcMutilDt}
\alias{ProcMutilDt}
\title{Process Multi Data}
\usage{
ProcMutilDt(MultiData)
}
\arguments{
\item{MultiData}{: MultiSV Data object}
}
\value{
\code{ProcMutilDt}: Process MultiSV data
}
\description{
Process Multi Data
}
\examples{
\dontrun{
ProcMutilDt(MultiData)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\seealso{
ProcMutilDt
}

