% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WriteMultiSV}
\alias{WriteMultiSV}
\title{Output MultiSVs}
\usage{
WriteMultiSV(MultiData,File)
}
\arguments{
\item{MultiData}{: MultiSV Data object}

\item{File}{: path to output File e.g. MultiSV.gff}
}
\value{
\code{WriteMultiSV}: read configuaration File
}
\description{
Output MultiSVs
}
\examples{
\dontrun{
WriteMultiSV(MultiData,File)
}
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}

