% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MultiSVData}
\alias{MultiSVData}
\title{RD Data}
\description{
The data set is read depth (RD) from populations of pigs (SS10.2). RD from each
population in 1kb bins.
}
\author{
Khurram \enc{Maqbool}{Maqbool}
}
\references{
Rubin, C.-J., H.-J.Megens, A.Martinez Barrio, K.Maqbool, S. Sayyab,
D. Schwochow, C.Wang, Ö. Carlborg, P. Jern, C.B. Jørgensen, A.
L.Archibald, M. Fredholm, M. A. M. Groenen, and L.Andersson.
2012. Strong signatures of selection in the domestic pig genome.
Proc. Natl. Acad. Sci. USA 109:19529–19536, DOI: 10.1073/pnas.1217149109
}
\keyword{CNV}
\keyword{Deletion}
\keyword{Duplication}
\keyword{MultiSV}
\keyword{Population}

